+++ src/main/java/com/taxicalls/trip/repository/PassengerRepository.java
@@ -3,6 +3,6 @@ package com.taxicalls.trip.repository;
 import com.taxicalls.trip.model.Passenger;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PassengerRepository extends CrudRepository<Passenger, Integer> {
+public interface PassengerRepository extends CrudRepository<Passenger, Long> {
 
 }
