+++ src/main/java/com/taxicalls/trip/repository/TripRepository.java
@@ -3,6 +3,6 @@ package com.taxicalls.trip.repository;
 import com.taxicalls.trip.model.Trip;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TripRepository extends CrudRepository<Trip, Integer> {
+public interface TripRepository extends CrudRepository<Trip, Long> {
 
 }
